refactor out structs from md
