+++ src/tests/md.rs
@@ -1,3 +1,4 @@
+use crate::md::structs::{MessDetectorChar, MessDetectorCharFlags};
 use crate::md::*;
 use crate::utils::{decode, get_large_test_datasets};
 use encoding::DecoderTrap;
